@@ -10,9 +10,10 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.bumptech.glide.Glide;
+import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions;
 import com.example.pichainventory.R;
 import com.example.pichainventory.Models.Upload;
-import com.squareup.picasso.Picasso;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -48,7 +49,36 @@ public class StockAdapter extends RecyclerView.Adapter<StockAdapter.StockViewHol
     @Override
     public void onBindViewHolder(@NonNull StockViewHolder holder, int position) {
         Upload upload = mUploads.get(position);
-        holder.bind(upload);
+        
+        // Set item name
+        holder.itemLabel.setText(upload.getmName());
+        
+        // Set price
+        holder.SellingPEditText.setText(String.valueOf(upload.getmSp()));
+        
+        // Set units with stock status
+        holder.UnitEditText.setText(String.valueOf(upload.getmUnits()));
+        holder.UnitEditText.setActivated(upload.getmUnits() == 0); // This will trigger the red border
+        
+        // Load image with placeholder and error handling
+        if (upload.getmImageUrl() != null && !upload.getmImageUrl().isEmpty()) {
+            Glide.with(mContext)
+                .load(upload.getmImageUrl())
+                .placeholder(R.drawable.placeholder)
+                .error(R.drawable.placeholder)
+                .transition(DrawableTransitionOptions.withCrossFade())
+                .into(holder.profilePhoto);
+        } else {
+            // If no image URL is provided, show placeholder
+            holder.profilePhoto.setImageResource(R.drawable.placeholder);
+        }
+
+        // Set click listener
+        holder.itemView.setOnClickListener(v -> {
+            if (mListener != null) {
+                mListener.onItemClick(position, upload);
+            }
+        });
     }
 
     @Override
@@ -57,51 +87,32 @@ public class StockAdapter extends RecyclerView.Adapter<StockAdapter.StockViewHol
     }
 
     public class StockViewHolder extends RecyclerView.ViewHolder {
-        private ImageView profilePhoto;
-        private TextView nameTextView;
-        private TextView sPtextView;
+        ImageView profilePhoto;
+        TextView itemLabel;
+        TextView SellingPEditText;
+        TextView UnitEditText;
 
         public StockViewHolder(@NonNull View itemView) {
             super(itemView);
-
             profilePhoto = itemView.findViewById(R.id.profilePhoto);
-            nameTextView = itemView.findViewById(R.id.nameTextView);
-            sPtextView = itemView.findViewById(R.id.sPtextView);
-
-            itemView.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    int position = getAbsoluteAdapterPosition();
-                    if (position != RecyclerView.NO_POSITION && mListener != null) {
-                        Upload upload = mUploads.get(position);
-                        mListener.onItemClick(position, upload);
-                    }
-                }
-            });
-        }
-
-        public void bind(Upload upload) {
-            nameTextView.setText(upload.getmName());
-            sPtextView.setText(String.valueOf(upload.getmSp()));
-
-            Picasso.get()
-                    .load(upload.getmImageUrl())
-                    .placeholder(R.drawable.placeholder)
-                    .fit()
-                    .centerCrop()
-                    .into(profilePhoto);
+            itemLabel = itemView.findViewById(R.id.itemLabel);
+            SellingPEditText = itemView.findViewById(R.id.SellingPEditText);
+            UnitEditText = itemView.findViewById(R.id.UnitEditText);
         }
     }
+
     public void setData(List<Upload> uploads) {
         this.originalUploads = new ArrayList<>(uploads);
         this.mUploads = uploads;
         notifyDataSetChanged();
     }
+
     public void filterUploads(List<Upload> filteredUploads) {
         mUploads.clear();
         mUploads.addAll(filteredUploads);
         notifyDataSetChanged();
     }
+
     public void resetData() {
         if (originalUploads != null) {
             this.mUploads = new ArrayList<>(originalUploads);
